Remove unnecessary usage of #include <iostream>. Rename templated impls from .cc to -impl.h.
