+++ src/examples/ntester.cc
@@ -36,7 +36,7 @@
 #include "user.h"
 #include "fatal_assert.h"
 #include "pty_compat.h"
-#include "networktransport.cc"
+#include "networktransport-impl.h"
 #include "select.h"
 
 using namespace Network;
